reorganize `dids` package structure
