+++ dids/src/main/kotlin/web5/sdk/dids/DidResolvers.kt
@@ -1,6 +1,7 @@
 package web5.sdk.dids
 
 import foundation.identity.did.DID
+import web5.sdk.dids.methods.key.DidKey
 
 /**
  * Type alias for a DID resolver function.
